factored out common part of assembled and matrix-free diffusion tests
