+++ tests/CMakeLists.txt
@@ -94,10 +94,10 @@ add_mpi_test( MpiImportExportTest.cpp Import_Export_Tests "1;2;4" )
 add_mpi_test( MeshDistributionTest.cpp Mesh_Distribution "1;2;4" )
 add_mpi_test( NodeThroughputInfoTest.cpp Node_Throughput_Calculation "1;2;4" )
 add_mpi_test( SparsityGraphTest.cpp Sparsity_Graph "1;2;4" )
-add_mpi_test( Diffusion2DTest.cpp Solve_2D_diffusion_problem "1;2;4" )
-add_mpi_test( AdvectionPeriodic2DAssembled.cpp Solve_2D_Periodic_Advection_Problem_Assembled "1;2;4" )
-add_mpi_test( AdvectionPeriodic2DMF.cpp Solve_2D_Periodic_Advection_Problem_Matrix_Free "1;2;4" )
-add_mpi_test( MatrixFreeTests.cpp Matrix_Free "1;2;4" )
+add_mpi_test( AdvectionPeriodic2DAssembled.cpp Solve_2D_periodic_advection_problem_assembled "1;2;4" )
+add_mpi_test( AdvectionPeriodic2DMF.cpp Solve_2D_periodic_advection_problem_matrix_free "1;2;4" )
+add_mpi_test( Diffusion2DAssembledTest.cpp Solve_2D_diffusion_problem_assembled "1;2;4" )
+add_mpi_test( Diffusion2DMF.cpp Solve_2D_diffusion_problem_matrix_free "1;2;4" )
 add_mpi_test( DirichletBCTest.cpp Dirichlet_BC_Application "1;2;4" )
 add_mpi_test( PeriodicBCTest.cpp Periodic_BC_test "1;2;4;16" )
 add_mpi_test( EmptyPartitionTest.cpp Handle_Empty_Partition "4" )
